09/09/2019 evening push

+++ javaprograms/bootcamp/src/com/bridgelabz/functional/PowerOf2.java
@@ -5,11 +5,11 @@ public class PowerOf2 {
 	public static void main(String[] args) {
 		Scanner sc=new Scanner(System.in);
 		System.out.println("Enter the value of n");
-		long n=sc.nextLong();
+		int n=sc.nextInt();
 		
-		for(long i=0;i<=n;i++)
+		for(int i=0;i<=n;i++)
 		{
-			System.out.println(Math.pow(2, i));
+			System.out.println((int)Math.pow(2, i));
 		}
 		sc.close();
 	}
